Revert "Fi 972 refactor icrc 1 ledger documentation"
