--- docs/developer-docs/integrations/internet-identity/overview.md
@@ -1,9 +1,0 @@
-# Internet Identity
-
-Internet Identity is an authentication service for the Internet Computer. It uses a form of the WebAuthn API to allow users to register and authenticate without passwords. Instead, users can use TouchID, FaceID, and more, to authenticate with an application. Interacting with Internet Identity is as simple as opening the Internet Identity dashboard, https://identity.ic0.app. 
-
-- [Internet Identity dashboard](https://identity.ic0.app/).
-- [Internet Identity specification](https://internetcomputer.org/docs/current/references/ii-spec).
-- [Internet Identity GitHub repository](https://github.com/dfinity/internet-identity).
-- [Internet Identity alternative frontend origins](https://internetcomputer.org/docs/current/developer-docs/integrations/internet-identity/alternative-origins).
-- [Internet Identity integration](https://internetcomputer.org/docs/current/developer-docs/integrations/internet-identity/integrate-identity).
