--- roadmap/1_core-protocol/future/3tb-storage.md
@@ -1,10 +1,0 @@
----
-title: Support 3TB replicated storage
-links:
-  Forum Link:
-  Proposal:
-eta: 
-is_community: false
----
-
-This feature is the follow up feature to supporting 1TB of replicated storage. The goal is to support 3TB of replicated storage on subnets. Work on this feature includes protocol improvements and optimizations related to interacting with the replicated state on disk and when syncing it via state sync. It also includes work to make sure that node machines match the hardware requirements in terms of physical disk size etc.
