+++ sidebars.js
@@ -414,7 +414,7 @@ const sidebars = {
           label: "Internet Identity (II)",
           link: {
             type: "doc",
-            id: "developer-docs/integrations/internet-identity/overview",
+            id: "references/ii-spec",
           },
           items: [
             "developer-docs/integrations/internet-identity/alternative-origins",
@@ -462,8 +462,8 @@ const sidebars = {
           },
           items: [
             "developer-docs/integrations/ledger/collecting-dust",
-            "developer-docs/integrations/ledger/ledger-local-setup",
             "developer-docs/integrations/ledger/interact-with-ledger",
+            "developer-docs/integrations/ledger/ledger-local-setup",
             "developer-docs/integrations/ledger/icp-index-local-setup",
           ],
         },
@@ -474,11 +474,7 @@ const sidebars = {
             type: "doc",
             id: "developer-docs/integrations/icrc-1/index",
           },
-          items: [
-            "developer-docs/integrations/icrc-1/icrc1-ledger-setup",
-            "developer-docs/integrations/icrc-1/interact-with-ICRC-1-ledger",
-          "developer-docs/integrations/icrc-1/icrc1-index-setup"
-        ],
+          items: ["developer-docs/integrations/icrc-1/icrc1-ledger-setup"],
         },
         {
           type: "doc",
