--- src/components/Community/Globe.tsx
@@ -1,58 +1,0 @@
-import createGlobe from "cobe";
-import React, { useEffect, useRef } from "react";
-import { Hub } from "./Hubs";
-
-const Globe: React.FC<{
-  hubs: Hub[];
-  className?: string;
-}> = ({ className, hubs }) => {
-  const canvasRef = useRef<HTMLCanvasElement>();
-
-  useEffect(() => {
-    let phi = 0;
-
-    canvasRef.current.width =
-      canvasRef.current.clientWidth * window.devicePixelRatio;
-    canvasRef.current.height =
-      canvasRef.current.clientHeight * window.devicePixelRatio;
-
-    const globe = createGlobe(canvasRef.current, {
-      devicePixelRatio: window.devicePixelRatio,
-      width: canvasRef.current.width,
-      height: canvasRef.current.height,
-      phi: 0,
-      theta: 0.42,
-      dark: 1,
-      diffuse: 1.75,
-      mapSamples: 43000,
-      mapBrightness: 7.3,
-      baseColor: [0.23, 0.0, 0.72], // #3b00b9 with each byte represented as 0..1
-      markerColor: [1, 1, 1],
-      // glowColor: [0.23, 0.0, 0.72],
-      glowColor: [0.8, 0.8, 0.8],
-      opacity: 0.41,
-      // longitude latitude
-      // { location: [37.7595, -122.4367], size: 0.03 },
-      markers: hubs.map((hub) => ({
-        location: hub.coordinates,
-        size: 0.03,
-      })),
-      onRender: (state) => {
-        // Called on every animation frame.
-        // `state` will be an empty object, return updated params.
-        state.phi = phi;
-        phi += 0.0025;
-        state.width = canvasRef.current.clientWidth * window.devicePixelRatio;
-        state.height = canvasRef.current.clientHeight * window.devicePixelRatio;
-      },
-    });
-
-    return () => {
-      globe.destroy();
-    };
-  }, []);
-
-  return <canvas className={className} ref={canvasRef}></canvas>;
-};
-
-export default Globe;
