--- src/components/Community/gallery-images.ts
@@ -1,37 +1,0 @@
-import { GalleryItem } from "./Gallery";
-
-export const communityGallery: GalleryItem[] = [
-  { image: "/img/community/image-000.webp" },
-  { image: "/img/community/image-001.webp" },
-  { image: "/img/community/image-002.webp" },
-  { image: "/img/community/image-003.webp" },
-  { image: "/img/community/image-004.webp" },
-  { image: "/img/community/image-005.webp" },
-  { image: "/img/community/image-006.webp" },
-  { image: "/img/community/image-007.webp" },
-  { image: "/img/community/image-008.webp" },
-  { image: "/img/community/image-009.webp" },
-  { image: "/img/community/image-010.webp" },
-  { image: "/img/community/image-011.webp" },
-  { image: "/img/community/image-012.webp" },
-  { image: "/img/community/image-013.webp" },
-  { image: "/img/community/image-014.webp" },
-  { image: "/img/community/image-015.webp" },
-  { image: "/img/community/image-016.webp" },
-  { image: "/img/community/image-017.webp" },
-  { image: "/img/community/image-018.webp" },
-  { image: "/img/community/image-019.webp" },
-  { image: "/img/community/image-020.webp" },
-
-  { image: "/img/community/image-021.webp" },
-  { image: "/img/community/image-022.webp" },
-  { image: "/img/community/image-023.webp" },
-  { image: "/img/community/image-024.webp" },
-  { image: "/img/community/image-025.webp" },
-  { image: "/img/community/image-026.webp" },
-  { image: "/img/community/image-027.webp" },
-  { image: "/img/community/image-028.webp" },
-  { image: "/img/community/image-029.webp" },
-  { image: "/img/community/image-030.webp" },
-  { image: "/img/community/image-031.webp" },
-];
